\name{data.eeg}
\alias{data.eeg}
\docType{data}
\title{
Example eeg data
}
\description{
data.eeg is an example of signal features extracted from eeg data
}
\usage{data(data.eeg)}
\format{
  The format is:
 chr "data.eeg"
}
\source{
The data was extracted as a subset from an study 
}
\examples{
data(data.eeg)
}
\keyword{datasets}