\name{data.labels}
\alias{data.labels}
\docType{data}
\title{
Example eeg data
}
\description{
Example labels for a dataset
}
\usage{data(data.labels)}
\format{
  The format is:
 chr "data.labels"
}
\source{
The data was extracted as a subset from an study 
}
\examples{
data(data.labels)
}
\keyword{datasets}